requirements.txt
pygame~=2
numpy~= 1.26
pillow~=10
pydantic
